+++ tests/v0x01/test_asynchronous/test_port_status.py
@@ -6,7 +6,7 @@ from pyof.v0x01.common.phy_port import (PhyPort, PortConfig, PortFeatures,
                                         PortState)
 from pyof.v0x01.foundation.base import OFP_MAX_PORT_NAME_LEN
 from pyof.v0x01.foundation.basic_types import HWAddress
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestPortStatus(TestStruct):
