+++ tests/v0x01/test_controller2switch/test_barrier_reply.py
@@ -1,6 +1,6 @@
 """Barrier reply message tests."""
 from pyof.v0x01.controller2switch.barrier_reply import BarrierReply
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestBarrierReply(TestStruct):
