+++ tests/v0x01/test_controller2switch/test_barrier_request.py
@@ -1,6 +1,6 @@
 """Barrier request message tests."""
 from pyof.v0x01.controller2switch.barrier_request import BarrierRequest
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestBarrierRequest(TestStruct):
