+++ tests/v0x01/test_controller2switch/test_features_reply.py
@@ -2,7 +2,7 @@
 from pyof.v0x01.common.phy_port import PhyPort, PortConfig, PortState
 from pyof.v0x01.controller2switch.features_reply import FeaturesReply
 from pyof.v0x01.foundation.basic_types import HWAddress
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestFeaturesReply(TestStruct):
