+++ tests/v0x01/test_controller2switch/test_features_request.py
@@ -1,6 +1,6 @@
 """Feature request message tests."""
 from pyof.v0x01.controller2switch.features_request import FeaturesRequest
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestFeaturesRequest(TestStruct):
