+++ tests/v0x01/test_controller2switch/test_packet_out.py
@@ -4,7 +4,7 @@ import unittest
 from pyof.v0x01.common.phy_port import Port
 from pyof.v0x01.controller2switch.packet_out import PacketOut
 from pyof.v0x01.foundation.exceptions import ValidationError
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestPacketOut(TestStruct):
