+++ tests/v0x01/test_controller2switch/test_set_config.py
@@ -1,7 +1,7 @@
 """Set Config message tests."""
 from pyof.v0x01.controller2switch.common import ConfigFlags
 from pyof.v0x01.controller2switch.set_config import SetConfig
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestSetConfig(TestStruct):
