+++ tests/v0x01/test_symmetric/test_echo_request.py
@@ -1,6 +1,6 @@
 """Echo request message tests."""
 from pyof.v0x01.symmetric.echo_request import EchoRequest
-from tests.teststruct import TestStruct
+from tests.test_struct import TestStruct
 
 
 class TestEchoRequest(TestStruct):
